Issue #12458 - Move get bundle logic to app manager
