+++ profile/modules/apps/os_widgets/os_widgets.info.yml
@@ -11,3 +11,4 @@ dependencies:
   - "drupal:os_mailchimp"
   - "drupal:paragraphs"
   - "drupal:slick"
+  - "drupal:vsite"
