+++ profile/modules/cp/modules/cp_taxonomy/cp_taxonomy.services.yml
@@ -1,7 +1,7 @@
 services:
   cp.taxonomy.helper:
     class: Drupal\cp_taxonomy\CpTaxonomyHelper
-    arguments: ['@config.factory', '@entity_type.manager', '@entity_type.bundle.info', '@vsite.context_manager', '@os_widgets.context', '@vsite.app.manager']
+    arguments: ['@config.factory', '@entity_type.manager', '@entity_type.bundle.info', '@vsite.context_manager']
   cp.taxonomy.list_subscriber:
     class: Drupal\cp_taxonomy\EventSubscriber\CpTaxonomyListingSubscriber
     tags:
