+++ profile/modules/cp/modules/cp_taxonomy/src/CpTaxonomyHelperInterface.php
@@ -110,12 +110,4 @@ interface CpTaxonomyHelperInterface {
    */
   public function getOptionsTree(string $vid): array;
 
-  /**
-   * Get all bundles from active apps.
-   *
-   * @return array
-   *   Return array of bundle keys.
-   */
-  public function getBundlesFromApps(): array;
-
 }
