+++ profile/modules/vsite/modules/vsite_privacy/vsite_privacy.info.yml
@@ -4,5 +4,5 @@ core: 8.x
 type: module
 version: '1.0'
 dependencies:
-  - drupal:vsite
-  - drupal:robotstxt
+  - "drupal:vsite"
+  - "drupal:robotstxt"
