+++ profile/modules/vsite/vsite.info.yml
@@ -2,5 +2,6 @@ name: VSite
 type: module
 core: 8.x
 dependencies:
-  - group
-  - purl
+  - drupal:group
+  - drupal:purl
+  - drupal:os_widgets
